NXP-25036: remove 'application/json+esentity' content type; move JsonESDocumentWriter to nuxeo-elasticsearch-core
