+++ nuxeo-features/nuxeo-elasticsearch/nuxeo-elasticsearch-automation/pom.xml
@@ -28,6 +28,10 @@
       <artifactId>nuxeo-automation-core</artifactId>
     </dependency>
     <dependency>
+      <groupId>org.nuxeo.ecm.platform</groupId>
+      <artifactId>nuxeo-platform-audit-api</artifactId>
+    </dependency>
+    <dependency>
       <groupId>org.nuxeo.elasticsearch</groupId>
       <artifactId>nuxeo-elasticsearch-core</artifactId>
     </dependency>
