+++ nuxeo-features/nuxeo-elasticsearch/nuxeo-elasticsearch-core/src/main/resources/OSGI-INF/elasticsearch-service-contrib.xml
@@ -2,7 +2,7 @@
 <component name="org.nuxeo.elasticsearch.ElasticSearchComponent.contrib" version="1.0">
 
   <extension point="elasticSearchDocWriter" target="org.nuxeo.elasticsearch.ElasticSearchComponent">
-    <writer class="org.nuxeo.ecm.automation.jaxrs.io.documents.JsonESDocumentWriter" />
+    <writer class="org.nuxeo.elasticsearch.io.JsonESDocumentWriter" />
   </extension>
 
 </component>
